% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getConditions.R
\name{getConditions}
\alias{getConditions}
\title{Pulls conditions available for a organism}
\usage{
getConditions(OrgID = NULL, group = NULL)
}
\arguments{
\item{OrgID}{A valid fitness browser organism identifier (see getOrganismCodes())}

\item{group}{(Optional) Show all genes with specific phenotypes for this condition}
}
\value{
A table with all conditions available for OrgID or a table with all specific phenotypes for a given condition group (if group is not null)
}
\description{
Pulls conditions available for a organism
}
