% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOrganismCodes.R
\name{getOrganismCodes}
\alias{getOrganismCodes}
\title{Retrieves organisms available in the Fitness Browser database}
\usage{
getOrganismCodes(filter = NULL)
}
\arguments{
\item{filter}{(Optional) A string to be used for subsetting the organism list}
}
\value{
A two-column table with organism/strain names and their corresponding IDs
}
\description{
Retrieves organisms available in the Fitness Browser database
}
