% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchFitnessBrowser.R
\name{searchFitnessBrowser}
\alias{searchFitnessBrowser}
\title{Searches a gene (or character vector of gene IDs) in the fitness browser}
\usage{
searchFitnessBrowser(gene = NULL, OrgID = NULL, filter = NULL)
}
\arguments{
\item{gene}{(Required) A string (or a character vector) with gene IDs to be searched in the fitness browser}

\item{OrgID}{(Required) The ID for the organism of the query}

\item{filter}{(Optional) a string used to filters the output table based on the conditions column

@returns A data frame with available fitness information for the queried genes}
}
\description{
Searches a gene (or character vector of gene IDs) in the fitness browser
}
